Fix decode of AC.

+++ Mangamix/jpeg/jtypes.h
@@ -14,7 +14,7 @@
 
 typedef uint8_t byte;
 
-typedef int8_t coeff_t;
+typedef int16_t coeff_t;
 
 
 #endif /* jtypes_h */
